+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # developing
 - [httplib: fix unstable unit test which use the httplib.org](https://github.com/beego/beego/pull/5232)
+- [remove adapter package](https://github.com/beego/beego/pull/5239)
 
 # v2.1.0
 - [unified gopkg.in/yaml version to v2](https://github.com/beego/beego/pull/5169)